:art: discard flexlayout, support display:flex
