+++ src/index.js
@@ -6,7 +6,6 @@ import Group from './group'
 import BaseNode from './basenode'
 import Path from './path'
 import Batch from './batch'
-import FlexLayout from './layout'
 import {registerNodeType, createNode} from './nodetype'
 import {Effects, Easings} from 'sprite-animator'
 import SvgPath from 'svg-path-to-canvas'
@@ -33,7 +32,6 @@ export {
   Path,
   Layer,
   Group,
-  FlexLayout,
   Effects,
   Easings,
   registerNodeType,
